// src/frontend/src/backend.d.ts
export interface NewRepeater {
    status: Status;
    dcsCode: string;
    coverageDescription: string;
    city: string;
    toneMode: string;
    submittedBy: string;
    offset: number;
    callSign: string;
    zipCode: string;
    state: string;
    sponsor: string;
    autopatchInfo: string;
    frequency: number;
    operationalNotes: string;
    linkInfo: string;
    ctcssTone: string;
}
// ...
export enum Status {
    active = "active",
    inactive = "inactive"
}